# apiDocumentation/swagger.yaml
openapi: 3.0.0
info:
  description: This is the ISE-Learning REST API
  version: 0.0.0
  title: ISE-Learning REST API
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
...
  '/v0/users/{userId}':  
    get:
      tags:
        - Users
      summary: Retreives an user
      operationId: getUser
      description: |
        Retrive an user's information.
      parameters:
        - in: path
          name: userId
          schema:
            type: integer
          required: true
          description: User's unique id.
      responses:
        '200':
          description: User
          content:
            application/json:
              schema:
                type: object
                properties:
                  payload:
                    $ref: '#/components/schemas/user'
    delete:
      tags:
        - Users
      summary: Deletes an user
      operationId: deleteUser
      description: |
        Delete an user.
      parameters:
        - in: path
          name: userId
          schema:
            type: integer
          required: true
          description: User's unique id.
      responses:
        '204':
          description: Successful deleted user.
          content:
            application/json:
              schema:
                type: object   
  '/v0/challenges':
    get:
      tags:
        - Challenges
      summary: Retreives all challenges
      operationId: getAllChallenges
      description: |
        Retrieves every challenge the user has read access to.
      parameters:
        - in: query
          name: privacy
          schema:
            type: string
            enum:
              - private
              - public
          required: false
          description: >-
            Optional filtering condition determining whether public or private
            challenges will be returned.
        - in: query
          name: tags
          schema:
            type: array
            items:
              type: string
          required: false
          description: Optional parameter filtering challenges by tags.
      responses:
        '200':
          description: List of challenges.
          content:
            application/json:
              schema:
                type: object
                properties:
                  payload:
                    type: array
                    description: List of challenges.
                    items:
                      $ref: '#/components/schemas/challenge'
    post:
      tags:
        - Challenges
      summary: Creates a challenge
      operationId: createsChallenge
      description: |
        Creates a challenge based on payload content.
      requestBody:
        description: Challenge to be created.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/challengeNoIdNoSolutionID'
      responses:
        '201':
          description: Successful challenge created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  payload:
                    $ref: '#/components/schemas/challenge'
  '/v0/challenges/{challengeId}':
    get:
      tags:
        - Challenges
      summary: Retreives a challenge
      operationId: getChallenge
      description: |
        Retrieves every challenge the user has read access to.
      parameters:
        - in: path
          name: challengeId
          schema:
            type: integer
          required: true
          description: Challenge unique id.
      responses:
        '200':
          description: Challenge
          content:
            application/json:
              schema:
                type: object
                properties:
                  payload:
                    type: object
                    $ref: '#/components/schemas/challenge'
    put:
      tags:
        - Challenges
      summary: Updates a challenge
      operationId: updateChallenge
      description: |
        Updates a challenge based on payload content and the id on the uri.
      parameters:
        - in: path
          name: challengeId
          schema:
            type: integer
          required: true
          description: Challenge unique id.
      requestBody:
        description: Challenge to be updated.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/challengeNoId'
      responses:
        '200':
          description: Successful updated challenge.
          content:
            application/json:
              schema:
                type: object
                properties:
                  payload:
                    $ref: '#/components/schemas/challenge'
    delete:
      tags:
        - Challenges
      summary: Deletes a challenge
      operationId: deleteChallenge
      description: |
        Delete a challenge based on the id on the uri.
      parameters:
        - in: path
          name: challengeId
          schema:
            type: integer
          required: true
          description: Challenge unique id.
      responses:
        '204':
          description: Successful deleted challenge.
          content:
            application/json:
              schema:
                type: object
  '/v0/challenges/users/{userId}':
    get:
      tags:
        - Challenges
      summary: Retreives all challenges of a given user
      operationId: getUserChallenges
      description: |
        Retrieves every challenge of a given user.
      parameters:
        - in: path
          name: userId
          schema:
            type: integer
          required: true
          description: User unique id.
        - in: query
          name: privacy
          schema:
            type: string
            enum:
              - private
              - public
          required: false
          description: >-
            Optional filtering condition determining whether public or private
            challenges will be returned.
        - in: query
          name: tags
          schema:
            type: array
            items:
              type: string
          required: false
          description: Optional parameter filtering challenges by tags.
      responses:
        '200':
          description: User challenge list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  payload:
                    type: array
                    description: List of challenges from a given user.
                    items:
                      $ref: '#/components/schemas/challenge'
  '/v0/challenges/questionnaires/{questionnaireId}':
    get:
      tags:
        - Challenges
      summary: Retreives all challenges from a given questionnaire
      operationId: getAllChallengesByQuestionnaireId
      description: |
        Retrieves every challenge from a questionnaire by its unique identifier.
      parameters:
        - in: path
          name: questionnaireId
          schema:
            type: integer
          required: true
          description: Questionnaire unique id.
      responses:
        '200':
          description: Challenge
          content:
            application/json:
              schema:
                type: object
                properties:
                  payload:
                    $ref: '#/components/schemas/challenge'
...
  '/v0/challenges/{challengeId}/tags':
    get:
      tags:
        - Challenge Tags
      summary: Retreives all challenge tags from a challenge
      operationId: getAllChallengeTagsByChallengeId
      description: |
        Retrieves every challenge tag associated to a specific challenge.
      parameters:
        - in: path
          name: challengeId
          schema:
            type: integer
          required: true
          description: Challenge unique id.
      responses:
        '200':
          description: List of tags.
          content:
            application/json:
              schema:
                type: object
                properties:
                  payload:
                    type: array
                    description: List of challenges.
                    items:
                      $ref: '#/components/schemas/challengeTag'
    post:
      tags:
        - Challenge Tags
      summary: Associates a tag with a Challenge
      operationId: createsChallengeTag
      description: |
        Associates a tag with a specific challenge.
      parameters:
        - in: path
          name: challengeId
          schema:
            type: integer
          required: true
          description: Challenge unique id.
      requestBody:
        description: Challenge tag to be created.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/challengeTagInputModel'
      responses:
        '201':
          description: Successful challenge tag created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  payload:
                    $ref: '#/components/schemas/challengeTag'
  '/v0/challenges/{challengeId}/tags/{tagId}':
    get:
      tags:
        - Challenge Tags
      summary: Retreives specific challenge tag from a challenge
      operationId: getAllChallengeTagsByChallengeIdAnTagText
      description: >
        Retrieves challenge tag associated to a specific challenge with a
        specific tag id.
      parameters:
        - in: path
          name: challengeId
          schema:
            type: integer
          required: true
          description: Challenge unique id.
        - in: path
          name: tagId
          schema:
            type: integer
          required: true
          description: Tag unique id.
      responses:
        '200':
          description: Specific challenge tag.
          content:
            application/json:
              schema:
                type: object
                properties:
                  payload:
                    type: object
                    description: Challenge tag.
                    items:
                      $ref: '#/components/schemas/challengeTag'
    delete:
      tags:
        - Challenge Tags
      summary: Delete specific challenge tag from a challenge
      operationId: deleteAllChallengeTagsByChallengeIdAnTagText
      description: >
        Delete challenge tag associated to a specific challenge with a specific
        tag id.
      parameters:
        - in: path
          name: challengeId
          schema:
            type: integer
          required: true
          description: Challenge unique id.
        - in: path
          name: tagId
          schema:
            type: integer
          required: true
          description: Tag unique id.
      responses:
        '204':
          description: Successful deleted challenge tag.
          content:
            application/json:
              schema:
                type: object
...
  '/v0/questionnaires/{questionnaireId}':
    get:
      tags:
        - Questionnaires
      summary: Retrieves a questionnaire
      operationId: getQuestionnaireById
      description: |
        Retrieves a single questionnaire by its unique identifer.
      parameters:
        - in: path
          name: questionnaireId
          schema:
            type: integer
          required: true
          description: Questionnaire unique identifer.
      responses:
        '200':
          description: Questionnaire
          content:
            application/json:
              schema:
                type: object
                properties:
                  payload:
                    $ref: '#/components/schemas/questionnaire'
    put:
      tags:
        - Questionnaires
      summary: Updates a questionnaire
      operationId: updateQuestionnaireById
      description: |
        Updates a questionnaire by its unique identifier.
      parameters:
        - in: path
          name: questionnaireId
          schema:
            type: integer
          required: true
          description: Questionnaire unique identifer.
      requestBody:
        description: Questionnaire to be updated.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/questionnaireModel'
      responses:
        '200':
          description: Successful updated questionnaire.
          content:
            application/json:
              schema:
                type: object
                properties:
                  payload:
                    $ref: '#/components/schemas/questionnaire'
    delete:
      tags:
        - Questionnaires
      summary: Deletes a questionnaire
      operationId: deleteQuestionnaireById
      description: |
        Delete a questionnaire by its unique identifier.
      parameters:
        - in: path
          name: questionnaireId
          schema:
            type: integer
          required: true
          description: Questionnaire unique identifer.
      responses:
        '204':
          description: Successful deleted questionnaire.
          content:
            application/json:
              schema:
                type: object
...
components:
  schemas:
...
    user:
      type: object
      properties:
        userId: 
          type: integer
          example: 523
          description: Unique id which uniquely identifies the user
        username:
          type: string
          example: "userOne"
          description: Unique string used to identify an user
        email:
          type: string
          example: "userOne@example.com"
          description: Email address which can be used to contact user
        name:
          type: string
          example: "John Doe"
          description: User's given name
...
    challenge:
      type: object
      required:
        - challengeID
        - challengeText
        - isPrivate
        - solutions
      properties:
        challengeID:
          type: integer
          example: 548
          description: Unique identifier of challenge.
        challengeText:
          type: string
          example: 'Implement Hello world, by printing Hello world into the console.'
          description: Test describing the challenge and its goal.
        isPrivate:
          type: boolean
          example: true
          description: >-
            Boolean value indicating whether the challenge can only be seen by
            its creator
        solutions:
          type: array
          description: List of challenges.
          items:
            $ref: '#/components/schemas/solution'
    challengeNoId:
      type: object
      required:
        - challengeText
        - isPrivate
        - solutions
      properties:
        challengeText:
          type: string
          example: 'Implement Hello world, by printing Hello world into the console.'
          description: Test describing the challenge and its goal.
        isPrivate:
          type: boolean
          example: true
          description: >-
            Boolean value indicating whether the challenge can only be seen by
            its creator
        solutions:
          type: array
          description: List of challenges.
          items:
            $ref: '#/components/schemas/solution'
    challengeNoIdNoSolutionID:
      type: object
      required:
        - challengeText
        - isPrivate
        - solutions
      properties:
        challengeText:
          type: string
          example: 'Implement Hello world, by printing Hello world into the console.'
          description: Test describing the challenge and its goal.
        isPrivate:
          type: boolean
          example: true
          description: >-
            Boolean value indicating whether the challenge can only be seen by
            its creator
        solutions:
          type: array
          description: List of challenges.
          items:
            $ref: '#/components/schemas/solutionNoID'
    solution:
      type: object
      required:
        - solutionId
        - codeLanguage
        - challengeCode
        - solutionCode
        - unitTests
      properties:
        solutionId:
          type: integer
          example: 79884
          description: Unique id to which uniquely identifies the solution.
        codeLanguage:
          type: string
          example: java
          description: Text specifying in which language the solution is.
        challengeCode:
          type: string
          example: 'public class code { public static void main(String[] args) { }}'
          description: Default code of the code presented to user.
        solutionCode:
          type: string
          example: >-
            public class code { public static void main(String[] args) {
            System.out.println(\"Hello World\");}}
          description: Solution of the challenge.
        unitTests:
          type: string
          example: >-
            import org.junit.Assert;import org.junit.Test;public class unitTests
            {@Test public void TestCase1() {Assert.assertTrue(true);}}
          description: Unit tests to which the the user submitted will be tested against.
    solutionNoID:
      type: object
      required:
        - codeLanguage
        - challengeCode
        - solutionCode
        - unitTests
      properties:
        codeLanguage:
          type: string
          example: java
          description: Text specifying in which language the solution is.
        challengeCode:
          type: string
          example: 'public class code { public static void main(String[] args) { }}'
          description: Default code of the code presented to user.
        solutionCode:
          type: string
          example: >-
            public class code { public static void main(String[] args) {
            System.out.println(\"Hello World\");}}
          description: Solution of the challenge.
        unitTests:
          type: string
          example: >-
            import org.junit.Assert;import org.junit.Test;public class unitTests
            {@Test public void TestCase1() {Assert.assertTrue(true);}}
          description: Unit tests to which the the user submitted will be tested against.
...
    challengeTag:
      type: object
      required:
        - challengeTagId
        - challengeId
        - tag
      properties:
        challengeTagId:
          type: integer
          example: 55874
          description: Unique challenge tag id.
        challengeId:
          type: integer
          example: 55874
          description: Unique challenge id to which the tag belongs to.
        tag:
          $ref: '#/components/schemas/tag'
    tag:
      type: object
      required:
        - tagId
        - tag
      properties:
        tagId:
          type: integer
          example: 55874
          description: Unique tag id.
        tag:
          type: string
          example: Array
          description: Unique tag text.
    challengeTagInputModel:
      type: object
      required:
        - tag
      properties:
        tag:
          type: string
          example: Array
          description: Unique tag text.
          
##################################
#         QUESTIONNAIRES         #
##################################
    questionnaire:
      type: object
      required:
        - questionnaireId
        - description
        - timer
        - creatorId
        - creationDate
      properties:
        questionnaireId:
          type: integer
          example: 21313
          description: questionnaire unique identifier.
        description:
          type: string
          example: Questionnaire for course PS-1920SV
          description: Text describing the questionnaire.
        timer:
          type: number
          example: 360000
          description: >-
            Time in milliseconds which an user has to solve the questionnaire.
            If value is NULL has no time limite.
        creatorId:
          type: integer
          example: 44598
          description: Unique user identifier of user who created the questionnaire.
        creationDate:
          type: string
          example: '2020/05/12'
          description: questionnaire's creation date in yyyy-MM-dd format
    questionnaireModel:
      type: object
      required:
        - description
        - timer
        - creatorId
      properties:
        description:
          type: string
          example: Questionnaire for course PS-1920SV
          description: Text describing the questionnaire.
        timer:
          type: number
          example: 360000
          description: >-
            Time in milliseconds which an user has to solve the questionnaire.
            If value is NULL has no time limite.
        creatorId:
          type: integer
          example: 44598
          description: Unique user identifier of user who created the questionnaire.
...